+++ mido/__init__.py
@@ -48,13 +48,13 @@ Getting started:
     >>> import mido
     >>> m = mido.Message('note_on', note=60, velocity=64)
     >>> m
-    <note_on message channel=0, note=60, velocity=64, time=0>
+    <message note_on channel=0, note=60, velocity=64, time=0>
     >>> m.type
     'note_on'
     >>> m.channel = 6
     >>> m.note = 19
     >>> m.copy(velocity=120)
-    <note_on message channel=0, note=60, velocity=64, time=0>
+    <message note_on channel=0, note=60, velocity=64, time=0>
     >>> s = mido.Message('sysex', data=[byte for byte in range(5)])
     >>> s.data
     (0, 1, 2, 3, 4)